Removed a warning caused by static_cast switched back to the original codes C style cast. Since it was ok for that code it really ought to be fine. Removed a typo in defferredmultisampling.cpp.

+++ base/vulkanexamplebase.cpp
@@ -1195,10 +1195,10 @@ void VulkanExampleBase::handleMessages(HWND hWnd, UINT uMsg, WPARAM wParam, LPAR
 		ValidateRect(window, NULL);
 		break;
 	case WM_KEYDOWN:
-		handleKeyEvent(convertRawKey(wParam), true, static_cast<uint32_t>(wParam));
+		handleKeyEvent(convertRawKey((uint32_t)wParam), true, (uint32_t)(wParam));
 		break;
 	case WM_KEYUP:
-		handleKeyEvent(convertRawKey(wParam), false, static_cast<uint32_t>(wParam));
+		handleKeyEvent(convertRawKey((uint32_t)wParam), false, (uint32_t)(wParam));
 		break;
 	case WM_RBUTTONDOWN:
 	case WM_LBUTTONDOWN:
